requirements.txt
tensorflow
streamlit
pillow
numpy
requests